Rename plural pages to singular
